02-PROOFREADER/pyproject.toml
[project]
name = "02-proofreader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "asgiref==3.8.1",
    "backoff==2.2.1",
    "bcrypt==4.2.1",
    "build==1.2.2.post1",
    "cachetools==5.5.0",
    "certifi==2024.12.14",
    "charset-normalizer==3.4.1",
    "chroma-hnswlib==0.7.6",
    "chromadb>=0.6.3",
    "click==8.1.8",
    "coloredlogs==15.0.1",
    "deprecated==1.2.15",
    "durationpy==0.9",
    "fastapi==0.115.6",
    "filelock==3.16.1",
    "flatbuffers==24.12.23",
    "fsspec==2024.12.0",
    "google-auth==2.37.0",
    "googleapis-common-protos==1.66.0",
    "grpcio==1.69.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "huggingface-hub==0.27.1",
    "humanfriendly==10.0",
    "idna==3.10",
    "importlib-metadata==8.5.0",
    "importlib-resources==6.5.2",
    "kubernetes==31.0.0",
    "langchain>=0.3.20",
    "langchain-chroma>=0.2.2",
    "langchain-community>=0.3.19",
    "langchain-core>=0.3.43",
    "langchain-groq>=0.2.5",
    "langchain-text-splitters>=0.3.6",
    "markdown-it-py==3.0.0",
    "mdurl==0.1.2",
    "mmh3==5.0.1",
    "monotonic==1.6",
    "mpmath==1.3.0",
    "numpy>=1.26.4",
    "oauthlib==3.2.2",
    "onnxruntime==1.20.1",
    "opentelemetry-api==1.29.0",
    "opentelemetry-exporter-otlp-proto-common==1.29.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.29.0",
    "opentelemetry-instrumentation==0.50b0",
    "opentelemetry-instrumentation-asgi==0.50b0",
    "opentelemetry-instrumentation-fastapi==0.50b0",
    "opentelemetry-proto==1.29.0",
    "opentelemetry-sdk==1.29.0",
    "opentelemetry-semantic-conventions==0.50b0",
    "opentelemetry-util-http==0.50b0",
    "orjson==3.10.14",
    "overrides==7.7.0",
    "packaging==24.2",
    "posthog==3.7.5",
    "protobuf==5.29.3",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.1",
    "pydantic==2.10.5",
    "pydantic-core==2.27.2",
    "pygments==2.19.1",
    "pymupdf>=1.25.3",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv>=1.0.1",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "rich>=13.9.4",
    "rsa==4.9",
    "safetensors==0.5.2",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "starlette==0.41.3",
    "streamlit>=1.43.1",
    "sympy==1.13.3",
    "tenacity==9.0.0",
    "tokenizers==0.21.0",
    "tqdm>=4.67.1",
    "transformers>=4.49.0",
    "typer==0.15.1",
    "types-aiofiles>=24.1.0.20241221",
    "types-tqdm>=4.67.0.20250301",
    "typing-extensions==4.12.2",
    "urllib3==2.3.0",
    "uvicorn==0.34.0",
    "uvloop==0.21.0",
    "watchdog>=6.0.0",
    "watchfiles==1.0.3",
    "websocket-client==1.8.0",
    "websockets==14.1",
    "wrapt==1.17.0",
    "zipp==3.21.0",
]
